pyproject.toml
[project]
name = "image"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pillow>=12.1.1",
]
